Create a new base class for migrated payloads
Payloads that migrated on DBus will use pretty much the same implementation
of the payload class, so let's create one from the RPM OSTree payload class.
